Refactor packages to match app architecture
