+++ app/src/main/java/app/wefridge/wefridge/application/WeFridgeFirebaseMessagingService.kt
@@ -1,4 +1,4 @@
-package app.wefridge.wefridge
+package app.wefridge.wefridge.application
 
 
 import android.app.NotificationChannel
@@ -9,6 +9,7 @@ import android.content.Intent
 import android.os.Build
 import android.util.Log
 import androidx.core.app.NotificationCompat
+import app.wefridge.wefridge.R
 import com.google.firebase.messaging.FirebaseMessagingService
 import com.google.firebase.messaging.RemoteMessage
