+++ app/src/main/java/app/wefridge/wefridge/application/exceptions/InternetUnavailableException.kt
@@ -1,4 +1,4 @@
-package app.wefridge.wefridge.exceptions
+package app.wefridge.wefridge.application.exceptions
 
 class InternetUnavailableException: Exception {
     constructor() : super("Internet unavailable")
