+++ app/src/main/java/app/wefridge/wefridge/application/exceptions/ItemIsSharedWithoutContactEmailException.kt
@@ -1,4 +1,4 @@
-package app.wefridge.wefridge.exceptions
+package app.wefridge.wefridge.application.exceptions
 
 class ItemIsSharedWithoutContactEmailException: Exception {
     constructor() : super("The Item was set to 'is shared' though without a contact email.")
