+++ app/src/main/java/app/wefridge/wefridge/application/model/AlertDialogs.kt
@@ -1,4 +1,4 @@
-package app.wefridge.wefridge.model
+package app.wefridge.wefridge.application.model
 
 import android.app.AlertDialog
 import android.content.Context
