+++ app/src/main/java/app/wefridge/wefridge/presentation/EditItemTextInputLayout.kt
@@ -1,4 +1,4 @@
-package app.wefridge.wefridge
+package app.wefridge.wefridge.presentation
 
 import android.R
 import android.content.Context
