+++ app/src/main/java/app/wefridge/wefridge/presentation/EndlessRecyclerOnScrollListener.kt
@@ -1,4 +1,4 @@
-package app.wefridge.wefridge
+package app.wefridge.wefridge.presentation
 
 import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
