+++ app/src/main/res/layout/fragment_nerby_item_list.xml
@@ -6,7 +6,7 @@
     android:id="@+id/swipe"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".NearbyItemFragment">
+    tools:context=".presentation.NearbyItemFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/list"
