+++ app/src/main/res/layout/fragment_pantry_list.xml
@@ -12,7 +12,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layoutManager="LinearLayoutManager"
-        tools:context=".PantryFragment"
+        tools:context=".presentation.PantryFragment"
         tools:listitem="@layout/fragment_pantry" />
